+++ assets/assets_config.json
@@ -1,0 +1,8 @@
+{
+   "images": [
+      "assets/images/RobotFramework_AIO.ico"
+   ],
+   "css": [],
+   "styles": [],
+   "html": []
+}
