+++ assets/styles/styles.css
@@ -1,0 +1,67 @@
+body {
+   background-color: #1e1e1e;
+   color: #d4d4d4;
+   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, sans-serif;
+   margin: 0;
+}
+.welcome-container {
+   background-color: #252526;
+   border-radius: 0;
+   padding: 1.5rem;
+   box-shadow: none;
+   min-height: 100vh;
+}
+@media (min-width: 576px) {
+   .welcome-container {
+       padding: 2rem;
+   }
+}
+@media (min-width: 992px) {
+   .welcome-container {
+       padding: 2rem 3rem;
+   }
+}
+.nav-tabs {
+   border-bottom: 1px solid #3c3c3c;
+}
+.nav-tabs .nav-link {
+   color: #d4d4d4;
+   border: none;
+   padding-bottom: 0.5rem;
+}
+.nav-tabs .nav-link.active {
+   color: #0078d4;
+   border-bottom: 2px solid #0078d4;
+   background-color: transparent;
+}
+.nav-tabs .nav-link:hover {
+   color: #ffffff;
+}
+.list-group-item {
+   background-color: #2d2d2d;
+   color: #d4d4d4;
+   border: 1px solid #3c3c3c;
+   transition: background-color 0.2s;
+}
+.list-group-item:hover {
+   background-color: #3c3c3c;
+}
+.list-group-item a {
+   color: #4fc1ff;
+   text-decoration: none;
+}
+.list-group-item a:hover {
+   text-decoration: underline;
+}
+.welcome-header img {
+   height: 40px;
+   margin-right: 1rem;
+}
+.welcome-header {
+   display: flex;
+   align-items: center;
+   margin-bottom: 2rem;
+}
+.tab-content {
+   padding: 1.5rem;
+}
